Two servers running and responding to requests
At this point, the two flask apps are running and responding to requests
on their respective ports. I'm not sure how to properly intercept and
handle the SIGINT with active threads. Right now it seems that sending
SIGINT will interrupt one of the flask apps and the other one will still
be running, requiring a second SIGINT to be sent to kill the process.

It also seems like Python threads can't really be killed by the main
thread by design. The general consensus is that running threads should
check for a condition that can be set by the main thread to exit
gracefully. That seems like it would need to be done inside the Flask
`run` method, which I don't want to try to fork and edit.

Will continue down this thread to see how to meet that last requirement.