Handled SIGINT to more gracefully end the app
Python threads are designed not to be just killed from the parent
thread. `thread.join()` will wait for the child thread to exit on its
own, but unless the thread itself was set up to accept a signal that it
checks in its execution loop, there's no other way to just shut it down
from the outside.

One approach could have been to create m;y own thread class that extends
the base `threading.Thread` class that accepts one of those signals. But
that thread would just be calling the flask `app.run()` method which on
its own doesn't have any built in way to stop it from the outside.

I did find that Flask requests can get the underlying server object from
the environment like so:

`request.environ.get('werkzeug.server.shutdown`)`

But as far as I could tell there wasn't a clean way to access that
outside of the context of processing a request. So one approach could
have been to have the SIGINT handler send a request to
0.0.0.0:4000/shutdown and 0.0.0.0:5000/shutdown and have the `/shutdown`
route shutdown the server.

What I ended up doing was just having the SIGINT handler call
`sys.exit(0)`. This doesn't exit the main thread when other threads are
still running unless those other threads are marked as daemons. That
seemed good enough for me for this exercise.